+++ packages/sn-filter-pane/src/components/FoldedListbox/FoldedListbox.tsx
@@ -1,3 +1,4 @@
+import { stardust } from '@nebula.js/stardust';
 import {
   Grid, styled, Tooltip, Typography,
 } from '@mui/material';
@@ -8,6 +9,7 @@ import type { IStores } from '../../store';
 import { COLLAPSED_HEIGHT } from '../ListboxGrid/distribute-resources';
 import DrillDown from './drillDown';
 import SelectionSegmentsIndicator from './SelectionSegmentsIndicator';
+import KEYS from '../keys';
 
 export interface FoldedListboxClickEvent {
   event: React.MouseEvent<HTMLDivElement>;
@@ -17,18 +19,32 @@ export interface FoldedListboxProps {
   resource: IListboxResource;
   onClick: ({ event, resource }: FoldedListboxClickEvent) => void;
   stores: IStores;
+  tabIndex?: number;
+  isInPopover?: boolean;
+}
+interface StyledGridProps {
+  constraints?: stardust.Constraints;
+  stardustTheme?: stardust.Theme;
+  containerRef: React.RefObject<HTMLDivElement>;
+}
+interface StyledDivProps {
+  isInPopover?: boolean;
 }
 
-export const FoldedListbox = ({ resource, onClick, stores }: FoldedListboxProps) => {
-  const fieldName = useFieldName(resource.layout);
-  const containerRef = useRef<HTMLDivElement>(null);
-  const {
-    constraints, stardustTheme, options, translator,
-  } = stores.store.getState();
-  const isRtl = options.direction === 'rtl';
-  const isDrillDown = resource.layout.qListObject.qDimensionInfo.qGrouping === 'H';
+const StyledDiv = styled('div', { shouldForwardProp: (p) => !['isInPopover'].includes(p as string) })<StyledDivProps>(
+  ({ isInPopover }) => ({
+    '&:focus': {
+      boxShadow: 'inset 0 0 0 2px #3F8AB3 !important',
+    },
+    '&:focus-visible': {
+      outline: 'none',
+    },
+    padding: isInPopover ? '2px' : undefined,
+  }),
+);
 
-  const StyledGrid = styled(Grid)(() => ({
+const StyledGrid = styled(Grid, { shouldForwardProp: (p) => !['constraints', 'stardustTheme', 'containerRef'].includes(p as string) })<StyledGridProps>(
+  ({ constraints, stardustTheme, containerRef }) => ({
     justifyContent: 'space-between',
     alignItems: 'flex-start',
     border: '1px solid #d9d9d9',
@@ -42,11 +58,55 @@ export const FoldedListbox = ({ resource, onClick, stores }: FoldedListboxProps)
     backgroundColor: stardustTheme?.getStyle('object', '', 'listBox.backgroundColor') ?? '#FFFFFF',
     color: stardustTheme?.getStyle('object', '', 'listBox.title.main.color') ?? '#404040',
     width: containerRef?.current?.clientWidth,
-  }));
+    '&:focus:not(:hover)': {
+      boxShadow: 'inset 0 0 0 2px #3F8AB3 !important',
+    },
+    '&:focus-visible': {
+      outline: 'none',
+    },
+  }),
+);
+
+export const FoldedListbox = ({
+  resource, onClick, stores, isInPopover,
+}: FoldedListboxProps) => {
+  const fieldName = useFieldName(resource.layout);
+  const containerRef = useRef<HTMLDivElement>(null);
+  const {
+    constraints, stardustTheme, options, translator,
+  } = stores.store.getState();
+  const isRtl = options.direction === 'rtl';
+  const isDrillDown = resource.layout.qListObject.qDimensionInfo.qGrouping === 'H';
+
+  const handleKeyDown = (event: React.KeyboardEvent) => {
+    const target = event.target as HTMLElement;
+    if (target.tagName === 'INPUT') {
+      return;
+    }
+    switch (event.key) {
+      case KEYS.ENTER:
+      // @ts-ignore
+        onClick({ event, resource });
+        break;
+      default:
+        return;
+    }
+
+    // Note: We should not stop propagation here since this folded listbox maybe deystroyed
+    // and we need to set focus to the listbox container after this.
+    event.preventDefault();
+  };
 
   return (
-    <div ref={containerRef}>
-      <StyledGrid container direction='column' data-testid={`collapsed-title-${fieldName}`} onClick={(event) => onClick({ event, resource })}>
+    <StyledDiv ref={containerRef} className="folded-listbox" onKeyDown={handleKeyDown} isInPopover={isInPopover}>
+      <StyledGrid
+        constraints={constraints}
+        stardustTheme={stardustTheme}
+        containerRef={containerRef}
+        container
+        direction='column'
+        data-testid={`collapsed-title-${fieldName}`}
+        onClick={(event) => onClick({ event, resource })}>
         <Grid container flexGrow={1} alignItems={'center'} sx={{ flexDirection: isRtl ? 'row-reverse' : 'row', flexWrap: 'nowrap' }} padding='0 8px'>
           {isDrillDown
             && <Tooltip title={translator?.get('Listbox.DrillDown')} enterDelay={2000}>
@@ -67,6 +127,6 @@ export const FoldedListbox = ({ resource, onClick, stores }: FoldedListboxProps)
           ></SelectionSegmentsIndicator>
         </Grid>
       </StyledGrid>
-    </div>
+    </StyledDiv>
   );
 };