+++ packages/sn-filter-pane/src/components/keys.ts
@@ -1,7 +1,10 @@
 const KEYS = {
+  ENTER: 'Enter',
   ESC: 'Escape',
   LEFT: 'ArrowLeft',
   RIGHT: 'ArrowRight',
+  UP: 'ArrowUp',
+  DOWN: 'ArrowDown',
 };
 
 export default KEYS;